NO-SNOW Add trace log to cancel query test
