+++ lib/logger.ts
@@ -29,6 +29,10 @@ export function getInstance() {
   // use the browser implementation of logger as the default implementation;
   // we do this so that unit tests don't fail when the modules they're testing
   // log messages
+  //
+  // TODO:
+  // BrowserLogger API doesn't match regular logger used by the driver, replace this with actual logger
+  // used in node when removing browser-related code.
   if (!instance) {
     instance = new BrowserLogger();
   }
